refactor(infra): keep retry attempt metadata internal
