+++ src/infra/retry.test.ts
@@ -1,7 +1,8 @@
 // Tests retry backoff timing and cancellation behavior.
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { MAX_TIMER_TIMEOUT_MS } from "../shared/number-coercion.js";
-import { getRetryAttemptErrors, resolveRetryConfig, retryAsync } from "./retry.js";
+import { getRetryAttemptErrors } from "./retry-attempt-errors.js";
+import { resolveRetryConfig, retryAsync } from "./retry.js";
 
 const randomMocks = vi.hoisted(() => ({
   generateSecureFraction: vi.fn(),
